+++ apps/hyperdrive-trading/src/ui/protocol/components/ProtocolLabel.tsx
