+++ apps/hyperdrive-trading/src/ui/protocol/components/ProtocolStatsBar.tsx
