+++ apps/hyperdrive-trading/src/ui/token/components/TokenLabel.tsx
